backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
Pillow==10.1.0
